requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-openai
langchain-community
sentence-transformers
faiss-cpu
pypdf